Reliably detect forks in containers
It's possible to have a process in a PID namespace that nevertheless
has tid == vtid (when the parent and nested PID counters happen to
overlap). In this case, we wrongly take the non-container path
while handling clone events and end up overwriting an unrelated
threadinfo (one that happened to have real pid == new child's vpid).

Fix by explicitly passing a flag meaning "the child process is in
a nested PID namespace" from the driver.